refactor(ENTESB-12218): Resolve utility god class Oas20ModelHelper
